pyproject.toml
[tool.poetry]
name = "baseproject"
version = "0.1.0"
description = "base project setup"
authors = ["RodolfoPena <rodolfo.p.ing@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.99.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.17"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
